### 2019-09-03 #1703 @jmansour
Hi @arfon.  The `paper.md` document isn't stored within the project repository, but is instead here:

https://github.com/underworldcode/joss/

It is unclear (unless I missed something) from the guidelines whether it was sufficient to have the document within a separate repo. I'd rather it remains in a separate repo, but it isn't a huge issue if you'd like it to be stored within the project repo.  If that is the case, will @whedon scan all branches? Or do you use the `version` to designate the branch?

Thanks
