### 2019-09-04 #1703 @arfon
Hi @jmansour, sorry this wasn't clear. If at all possible, we prefer the paper to be in the same repository as the submission software. Whedon can compile papers from branches other than the default one (usually `master`) so if you're able to place the paper in the https://github.com/underworldcode/underworld2 repo, that would be great.

BTW, I've slightly [tweaked our submission guidelines](https://github.com/openjournals/joss/commit/9e881a7606db2e1f3c84114ffafc2cfcbd670e5b) to hopefully make this more clear.
