### 2019-09-04 #1703 @arfon
> With regards to process, is it reasonable to create a branch to capture changes relating to the review, with a new project release created at the conclusion which can then be tied to the JOSS publication?

Yes, that would be great.
