### 2019-09-04 #1703 @jmansour
Ok, excellent. I'll add the document shortly.

Possible reviewers include Gabriele Morra (@gabersyd) and Sergio Zlotnik (@techas), although I believe Sergio has some experience using Underworld2 already. From the JOSS list of reviewers, I think any of the following might be appropriate: lheagy, whimian, daniellivingston, vidalgp, leouieda.
