### 2019-09-05 #1703 @leouieda
@jmansour thank you for the reviewer recommendations! They really help. I'll try to get someone with geodynamics experience and someone with Python/C FEM experience to balance things out.
