### 2019-09-05 #1703 @leouieda
:wave: Hi @florian-wagner @gabersyd would you like to review this submission for JOSS? Please see  [our documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information on what a JOSS review is like.
