### 2019-09-05 #1703 @florian-wagner
Hi @leouieda, thanks for consideration. Although I have been looking forward to a chance to contribute to JOSS, I am currently on job transition / moving and not available for reviews. Sorry.
