### 2019-09-05 #1703 @gabersyd
Hi, I have never interacted with JOSS so I am certainly curious and willing
to do my best to review it. I am very busy until Sep 19, but immediately
after that I can look at it and provide my feedback.

Gabriele

Gabriele Morra
<http://scholar.google.com/citations?sortby=pubdate&hl=en&user=DDuTppAAAAAJ&view_op=list_works>
, Department of Physics <http://physics.louisiana.edu>, School of
Geosciences <http://geos.louisiana.edu>, University of Louisiana at
Lafayette <http://www.louisiana.edu>

On Thu, Sep 5, 2019 at 7:55 AM Leonardo Uieda <notifications@github.com>
wrote:

> 👋 Hi @florian-wagner <https://github.com/florian-wagner> @gabersyd
> <https://github.com/gabersyd> would you like to review this submission
> for JOSS? Please see our documentation
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> for more
> information on what a JOSS review is like.
>
