### 2019-09-06 3c90993
Init commit for JOSS paper.
