+++ docs/joss/paper.bib
@@ -1,0 +1,102 @@
+@Article{MoresiEtAl2002,
+author="Moresi, L. and Dufour, F. and M{\"u}hlhaus, H.-B.",
+title="Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling",
+journal="Pure and Applied Geophysics",
+volume="159",
+number="10",
+pages="2335--2356",
+year="2002"
+}
+
+@Article{MoresiEtAl2003,
+author="Moresi, L. and Dufour, F. and M{\"u}hlhaus, H.-B.",
+title="A Lagrangian integration point finite element method for large deformation modeling of viscoelastic geomaterials",
+journal="Journal of Computational Physics",
+volume="184",
+number="2",
+pages="476--497",
+year="2003"
+}
+
+
+@article{MoresiEtAl2007,
+author = "L. Moresi and S. Quenette and V. Lemiale and C. Mériaux and B. Appelbe and H.-B. Mühlhaus",
+title = "Computational approaches to studying non-linear dynamics of the crust and mantle ",
+journal = "Physics of the Earth and Planetary Interiors ",
+volume = "163",
+number = "1–4",
+pages = "69 - 82",
+year = "2007",
+note = "Computational Challenges in the Earth Sciences "
+}
+
+@conference{Kluyver:2016aa,
+	Author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing},
+	Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+	Editor = {F. Loizides and B. Schmidt},
+	Organization = {IOS Press},
+	Pages = {87 - 90},
+	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
+	Year = {2016}
+}
+
+@article{Beucher2019UWGeodynamics,
+	journal = {Journal of Open Source Software},
+	doi = {10.21105/joss.01136},
+	issn = {2475-9066},
+	number = {36},
+	publisher = {The Open Journal},
+	title = {UWGeodynamics: A teaching and research tool for numerical geodynamic modelling},
+	url = {http://dx.doi.org/10.21105/joss.01136},
+	volume = {4},
+	author = {Beucher, Romain and Moresi, Louis and Giordani, Julian and Mansour, John and Sandiford, Dan and Farrington, Rebecca and Mondy, Luke and Mallard, Claire and Rey, Patrice and Duclaux, Guillaume and Kaluza, Owen and Laik, Arijit and Morón, Sara},
+	pages = {1136},
+	date = {2019-04-11},
+	year = {2019},
+	month = {4},
+	day = {11},
+}
+
+@misc{owen_kaluza_2019_3348451,
+  author       = {Owen Kaluza and
+                  Louis Moresi and
+                  John Mansour and
+                  David G Barnes},
+  title        = {OKaluza/LavaVu: v1.4.3},
+  month        = jul,
+  year         = 2019,
+  doi          = {10.5281/zenodo.3348451},
+  url          = {https://doi.org/10.5281/zenodo.3348451}
+}
+
+@Article{BeallEtAl2018,
+author="Beall, A. and Moresi, Cooper C.M.",
+title="Formation of cratonic lithosphere during the initiation of plate tectonics",
+journal="Geology",
+volume="46",
+number="6",
+pages="487--490",
+year="2018"
+}
+
[…]
