### 2019-09-06 62003bd
remove mirko for now as it's preventing compilation of document.
