### 2019-09-10 8cdf792
* Reorganised function debug info to minimise file system inpact: ** By default, only root proc generates info. Users can set    the UW_WORLD_FUNC_MESSAGES env variable to get all procs to    report ** Switch to alternate methods of obtaining info to minimise    `stat()` calls. ** Better heuristics for determining which part of the stack    to display. This minimises `stat()` calls, and also gives    better info for the user. ** Optionally disable info altogether using UW_NO_FUNC_MESSAGES    env variable. * Similarly, switched to better methods to minimise `stat()`   calls when using timing mo
[…]
