+++ underworld/mesh/_mesh.py
@@ -162,7 +162,7 @@ class FeMesh(_stgermain.StgCompoundComponent, function.FunctionInput):
         returned array is a proxy for all the *local* vertices, and it is
         provided as 1d list.
 
-        As these arrays are simply proxys to the underlying memory structures,
+        As these arrays are simply proxies to the underlying memory structures,
         no data copying is required.
 
         Note that this property returns a read-only numpy array as default. If
