### 2019-09-16 54605b9
For sphinx, updated mocked classes for pip changes.
