### 2019-09-27 c83d5f0
Improvements to othogonal slcn algorithm.
Better handling for inflow models with slcn:
- Improved detection/handling for off mesh interpolation positions.
- Spatial tolerances requirement for Bicubic interpolation.

Various coding improvments for speed and readability.

Updated thermomechanical models to test slcn.
