+++ underworld/libUnderworld/StgFEM/Utils/src/SemiLagrangianIntegrator.c
@@ -189,9 +189,7 @@ void _SemiLagrangianIntegrator_Initialise( void* slIntegrator, void* data ) {
    FeVariable*			feVariable;
    FeVariable*			feVarOld;
    FeVariable*			feVarStar;
-   unsigned			   lMeshSize;
-   unsigned			   field_i, node_i;
-   double				phi[3];
+   unsigned			   field_i;
 
    if(self->velocityField) Stg_Component_Initialise(self->velocityField, data, False);
    if(self->advectedField) Stg_Component_Initialise(self->advectedField, data, False);
@@ -240,7 +238,6 @@ void SemiLagrangianIntegrator_InitSolve( void* _self, void* _context ) {
    FeVariable*			feVariable;
    FeVariable*			feVarOld;
    FeVariable*			feVarStar;
-   double				dt			= self->context->dt;
    double            phi[3];
    unsigned			   lMeshSize;
    FeMesh*				mesh;
@@ -356,7 +353,7 @@ void SemiLagrangianIntegrator_UpdatePreviousVelocityField( void* _self, void* _c
 }
 
 /* cubic Lagrangian interpoation in 1-D */
-void InterpLagrange( double x, double* coords, double** values, unsigned numdofs, double* result ) {
+void InterpLagrange( double x, double* coords, double (*values)[3], unsigned numdofs, double* result ) {
    unsigned	node_i, dof_i;
    unsigned	otherIndices[3];
    unsigned	otherIndexCount, otherIndex_i;
@@ -394,24 +391,40 @@ Bool PeriodicUpdate( double* pos, double* min, double* max, unsigned dim, Bool i
   return False;
 }
 
-void BicubicInterpolator( FeVariable* feVariable, double* position, double* delta, unsigned* nNodes, double* result ) {
-   FeMesh*		feMesh			= feVariable->feMesh;
-   Index		elementIndex;
-   unsigned	nInc, *inc;
-   int		x_0, y_0, z_0;
+Bool BicubicInterpolator( FeVariable* feVariable, double* position, double* delta, unsigned* nNodes, double* result ) {
+  /* Calculated the BicubicInterpolation of the feVariable at position
+   *
+   * Input Args:
+   *   feVariable: the field to be interpolated at `position`.
+   *   position:   the position of interpolatation.
+   *   delta:      memory chunk of size = sizeof(double)*dim.
+   *   nNodes:     numbers of nodes in ijk representation.
+   *   results:    the interpolated value.
+   *
+   *
+   * Returns Values:
+   *  True: if interpolation successful
+   *  False: position is not in 'domain' of local processor
+   */
+
+   FeMesh*	feMesh			= feVariable->feMesh;
+   unsigned	nInc ;
+   int		ijk[3],*inc;
    int		x_i, y_i, z_i;
-   double		localMin[3], localMax[3];
-   Index		gNode_I, lNode_I;
-   double		px[4], py[4], pz[4];
-   unsigned	nDims			= Mesh_GetDimSize( feMesh );
+   double localMin[3], localMax[3];
+   Index  gNode_I, lNode_I, elementIndex;
+   double px[4], py[4], pz[4];
    unsigned	nodeIndex[4][4];
    unsigned	node_I3D[4][4][4];
-   unsigned	numdofs			= feVariable->dofLayout->dofCounts[0];
-   double**	ptsX			= Memory_Alloc_2DArray_Unnamed( double, 4, 3 );
-   double**	ptsY			= Memory_Alloc_2DArray_Unnamed( double, 4, 3 );
-   double**	ptsZ			= Memory_Alloc_2DArray_Unnamed( double, 4, 3 );
[…]
