### 2019-10-09 #1703 @leouieda
:wave: Hi @krischer would you be interested in reviewing this submission for JOSS? Here is a [link to the submitted paper](https://github.com/openjournals/joss-papers/blob/joss.01703/joss.01703/10.21105.joss.01703.pdf) and the [software repository](https://github.com/underworldcode/underworld2).
