### 2019-10-09 #1703 @krischer
Hi @leouieda,

thanks for the consideration and while I could review the Python part I'm not familiar with geodynamic simulations at all so I don't feel qualified for this particular review.

I think @gassmoeller for example would be much more qualified here.
