### 2019-10-10 #1797 @whedon
[REVIEW]: Underworld2: Python Geodynamics Modelling for Desktop, HPC and Cloud
**Submitting author:** @jmansour (<a href="http://orcid.org/0000-0001-5865-1664">john mansour</a>)
**Repository:** <a href="https://github.com/underworldcode/underworld2" target ="_blank">https://github.com/underworldcode/underworld2</a>
**Version:** v2.9.2b
