### 2019-10-10 #1797 @leouieda
👋 @jmansour @gassmoeller @gabersyd this is the review thread for the paper. All of our communications will happen here from now on.

We aim for reviews to be completed within about 4 weeks. Please let me know if any of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping me here (@leouieda) or email me privately if you have any questions/concerns.
