### 2019-10-11 underworldcode/underworld2#423 @gassmoeller
Missing License-LGPLv3.md file
underworld2/LICENSE.md mentions a License-LGPLv3.md file that contains a copy of the LGPL license, but I do not see such a file. Did I simply not find the file, or is it missing?

Related to openjournals/joss-reviews#1797
