### 2019-10-15 b3beb12
implemented suggestions after code review
