### 2019-10-15 36646ff
Improvements to othogonal slcn algorithm. (#421)
These fixes only concerns a known bug for inflow problems modelled with Q1 elements

Better handling for inflow models with slcn:
- Improved detection/handling for off mesh interpolation positions.
- Spatial tolerances requirement for Bicubic interpolation.

Various coding improvements for speed and readability.

Updated thermomechanical models to test slcn.

Implemented suggestions after code review
