### 2019-10-16 aef32b6
Revert "Add a Gitter chat badge to README.md"
