### 2019-10-16 c4ad5b9
Add missing declaration of DBL_EPSILON
