### 2019-10-16 ee38d17
Use Petsc machine epsilon const
