### 2019-10-16 32be9ee
Add missing declaration of DBL_EPSILON (#426)
* Add missing declaration of DBL_EPSILON

* Use Petsc machine epsilon const
