+++ underworld/libUnderworld/StgFEM/Utils/src/SemiLagrangianIntegrator.c
@@ -11,6 +11,7 @@
 #include <StGermain/StGermain.h>
 #include <StgDomain/StgDomain.h>
 #include <StgFEM/StgFEM.h>
+#include <petsc.h>
 
 #include "types.h"
 #include "SemiLagrangianIntegrator.h"
@@ -391,6 +392,7 @@ Bool PeriodicUpdate( double* pos, double* min, double* max, unsigned dim, Bool i
   return False;
 }
 
+
 Bool BicubicInterpolator( FeVariable* feVariable, double* position, double* delta, unsigned* nNodes, double* result ) {
   /* Calculated the BicubicInterpolation of the feVariable at position
    *
@@ -482,11 +484,11 @@ Bool BicubicInterpolator( FeVariable* feVariable, double* position, double* delt
       for( d_i=0; d_i<nDims; d_i++ ) {
         // if not in the left most cell 
         tmp = position[d_i] - (localMin[d_i]+delta[d_i]);
-        if( tmp > DBL_EPSILON ) ijk[d_i]--;
+        if( tmp > PETSC_MACHINE_EPSILON ) ijk[d_i]--;
         
         // if in the right most cell 
         tmp = position[d_i] - (localMax[d_i]-delta[d_i]);
-        if( tmp > DBL_EPSILON ) ijk[d_i]--;
+        if( tmp > PETSC_MACHINE_EPSILON ) ijk[d_i]--;
       }
 
    }
