### 2019-10-16 9045c24
* Fix segfault, which was due to recent rearrangements. * Tidy to remove warnings.
