### 2019-10-17 a226118
Fixes #423
* Add LGPLv3 text.
* Update `LICENSE.md`
* Update release guidelines with copyright reminder.
