### 2019-10-23 underworldcode/underworld2#428 @gassmoeller
Authors for joss paper
This might have been resolved internally, but just to be sure: In https://github.com/underworldcode/underworld2/blob/joss/docs/joss/paper.md you have the following comment:

Has this been resolved, i.e. is the author list of the paper up-to-date?

Related to openjournals/joss-reviews/issues/1797
