### 2019-10-23 underworldcode/underworld2#429 @gassmoeller
Status of Jenkins CI server?
Related to openjournals/joss-reviews#1797

I was trying to take a look at the Jenkins build status that is linked in README.md, but I only get connection timeouts for both master and development branches (http://130.56.252.251:32779/job/Underworld2/job/master/ and http://130.56.252.251:32779/job/Underworld2/job/development/). Is this a misunderstanding on my end, or is the CI server currently offline? I can take a look at the Jenkinsfile directly, so no need to rush to fix this, but without accessible results the CI system is not much help. For the future you could also think about integrating the results of Jenkins directly to your PRs (e.g. see the bottom of https://github.com/geodynamics/aspect/pull/3271). We have made the experience that immediate feedback about broken tests is a great motivation for users to fix their PRs (and it saves some time for PR reviewers).
