### 2019-10-23 9b446dc
Fix some typos. Mark up some todos.
