### 2019-10-24 #1797 @gassmoeller
@leouieda: I have a question about the following item on the reviewer checklist:

```
State of the field: Do the authors describe how this software compares to other commonly-used packages?
```

The authors do not explicitly describe this in the paper (although they do describe their main design principle of an easily usable library, which differs somewhat from many of the other software packages in our field), but I also have two problems with this item. First it is not mentioned on the https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss website, and neither on the https://joss.readthedocs.io/en/latest/review_criteria.html#review-criteria website. Thus, the authors were not asked to provide such a comparison. Second, this item can be a can of worms if you want to describe it thoroughly. I think this item should be revised somewhat to either only require a short / general comparison (including mentioning this expectation on the submission website), or it should be dropped from the review checklist. Do you think this is reasonable?
