### 2019-10-25 46a8433
New stencil based orth slcn approach.
* Evaluate the stencils using global ijk representation
( Not global coords as previously, which caused floating point
approximation problems in parallel)

* Precompute the stencil beginning and store the 3-tuple on a
MeshVariable
