+++ underworld/libUnderworld/StgFEM/Utils/src/SemiLagrangianIntegrator.c
@@ -392,6 +392,109 @@ Bool PeriodicUpdate( double* pos, double* min, double* max, unsigned dim, Bool i
   return False;
 }
 
+Bool New_BicubicInterpolator( FeVariable* feVariable, FeVariable* stencilField, double* position, unsigned* sizes, double* result ) {
+  /* Calculated the BicubicInterpolation of the feVariable at position
+   *
+   *
+   *
+   * Returns Values:
+   *  True: if interpolation successful
+   *  False: position is not in 'domain' of local processor
+   */
+
+   FeMesh*	feMesh = feVariable->feMesh;
+   unsigned	nInc ;
+   int    ijk[3];
+   int		x_i, y_i, z_i, *inc;
+   double localMin[3], localMax[3], double_ijk[3];
+   Index  gNode_I, lNode_I, elementIndex;
+   double px[4], py[4], pz[4];
+   unsigned	nodeIndex[4][4];
+   unsigned	node_I3D[4][4][4];
+   unsigned	nDims   = Mesh_GetDimSize( feMesh );
+   unsigned	numdofs = feVariable->dofLayout->dofCounts[0];
+   double   ptsX[4][3], ptsY[4][3], ptsZ[4][3];
+   Bool     inDomain  = True;
+
+   inDomain = Mesh_SearchElements( feMesh, position, &elementIndex ); // get the element id
+
+   if (!inDomain) return False;
+
+   FeMesh_GetElementNodes( feMesh, elementIndex, feVariable->inc ); // get the incidence graph (inc.) of nodes on the element
+   nInc = IArray_GetSize( feVariable->inc );                        // from inc. get the number of nodes on the element
+   inc = IArray_GetPtr( feVariable->inc );                          // get the node ids from inc.
+
+   FeVariable_GetValueAtNode( stencilField, inc[0], &(double_ijk[0]) );
+   ijk[0] = lround(double_ijk[0]);
+   ijk[1] = lround(double_ijk[1]);
+   ijk[2] = lround(double_ijk[2]);
+
+   /* interpolate using Lagrange's formula */
+   if( nDims == 2 ) {
+      for( y_i = 0; y_i < 4; y_i++ )
+         for( x_i = 0; x_i < 4; x_i++ ) {
+            gNode_I = ijk[0] + x_i + ( ijk[1] + y_i ) * sizes[0];
+            if( !Mesh_GlobalToDomain( feMesh, MT_VERTEX, gNode_I, &lNode_I ) ){
+              printf("Error in BicubicInterpolator(), trying to build an interpolation to position (%g, %g) using node %d, a non domain node, in interpolator.\n", position[0], position[1], gNode_I);
+              abort();
+            }
+            else
+               nodeIndex[x_i][y_i] = lNode_I;
+         }
+   }
+   else {
+      for( z_i = 0; z_i < 4; z_i++ )
+         for( y_i = 0; y_i < 4; y_i++ )
+            for( x_i = 0; x_i < 4; x_i++ ) {
+               gNode_I = ijk[0] + x_i + ( ijk[1] + y_i ) * sizes[0] + ( ijk[2] + z_i ) * sizes[0] * sizes[1];
+               if( !Mesh_GlobalToDomain( feMesh, MT_VERTEX, gNode_I, &lNode_I ) ) {
+                 printf("Error in BicubicInterpolator(), trying to build an interpolation to position (%g, %g, %g) using node %d, a non domain node, in interpolator.\n", position[0], position[1], position[2], gNode_I);
+                 abort();
+               }
+               else
[…]
