### 2019-10-28 025deea
Documentation update for Jupyter tokens. (#431)
* Documentation update for Jupyter tokens.
* Increment UW version number
* update docker jupyter config for token
* update CHANGES.md
