+++ underworld/__init__.py
@@ -59,7 +59,7 @@ import ctypes as _ctypes
 _oldflags = _sys.getdlopenflags()
 _sys.setdlopenflags( _oldflags | _ctypes.RTLD_GLOBAL )
 
-__version__ = "2.8.1b"
+__version__ = "2.8.2b"
 
 # squelch h5py/numpy future warnings
 import warnings as _warnings
