### 2019-10-29 c37a4da
Move lavavu build to base image.
