### 2019-10-30 030b3af
Documentation updates
