### 2019-10-30 #1797 @leouieda
> I think this item should be revised somewhat to either only require a short / general comparison (including mentioning this expectation on the submission website), or it should be dropped from the review checklist. Do you think this is reasonable?

Hi @gassmoeller that is probably a new item on the checklist. I hadn't seen it before. I'll bring this up with the JOSS editors and see what we can do about it. I agree that we should be very explicit about what we want from that statement. I imagine the idea was to provide a general comparison statement, not a thorough performance/accuracy/features comparison. Thanks for pointing this out!
