### 2019-10-30 #1797 @gassmoeller
I am satisfied with the current state of the software and paper, and think this is an important contribution. The authors have responded productively and fixed the small comments and objections I had. I am only waiting for underworldcode/underworld2#428 to be fixed (one of the authors is currently not on the paper, because he has no ORCID so far), and afterwards this is ready from my side. I particularly liked the easy distribution and online tutorials. Nice work!
