### 2019-10-31 #1797 @gabersyd
@leouieda, I need one more week to give a complete feedback.

Gabriele Morra
<http://scholar.google.com/citations?sortby=pubdate&hl=en&user=DDuTppAAAAAJ&view_op=list_works>
, Department of Physics <http://physics.louisiana.edu>, School of
Geosciences <http://geos.louisiana.edu>, University of Louisiana at
Lafayette <http://www.louisiana.edu>

On Wed, Oct 30, 2019 at 1:45 PM Rene Gassmoeller <notifications@github.com>
wrote:

> I am satisfied with the current state of the software and paper, and think
> this is an important contribution. The authors have responded productively
> and fixed the small comments and objections I had. I am only waiting for
> underworldcode/underworld2#428
> <https://github.com/underworldcode/underworld2/issues/428> to be fixed
> (one of the authors is currently not on the paper, because he has no ORCID
> so far), and afterwards this is ready from my side. I particularly liked
> the easy distribution and online tutorials. Nice work!
>
