### 2019-10-31 #1797 @jmansour
Hi @leouieda

Are the ORCID identifiers strictly necessary for submission?  One of our authors is somewhat privacy concerned and would rather not sign up for an ORCID.
