### 2019-10-31 #1797 @leouieda
@jmansour ORCIDs are encouraged but not strictly required if an author feels strongly about not providing one (our own [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) has an authors without an ORCID).
