### 2019-10-31 c85452d
Add missing "References" heading to JOSS paper
JOSS papers should end with a `# References` to separate the Acknowledgements
from the BibTeX reference list that is automatically inserted at the end.

openjournals/joss-reviews#1797
