### 2019-10-31 7c0c4df
Add missing "References" heading to JOSS paper (#435)
JOSS papers should end with a `# References` to separate the Acknowledgements
from the BibTeX reference list that is automatically inserted at the end.

openjournals/joss-reviews#1797
