### 2019-11-01 #1797 @jmansour
Thanks @leouieda

A further question if I may.  One author has since moved onto other employment and is no longer directly involved in the project.  I assume that their affiliation should represent their place of employment at the time of their contribution, but possibly their current employment should be also  included as means of contact.  Does JOSS have any particular preference in this situation?
