### 2019-11-01 31c66d4
Move to multistage builds: * lavavu now built in `base/Dockerfile_lavavu`. * `mpi`,`mpi4py`,`petsc`,`petsc4py`,`h5py` built in `base/Dockerfile_petsc` * Each new Dockerfile generates a self contained image which can   be executed to run only the respective stacks * `base/Dockerfile` adds minimal extra packages, and extracts 'artifacts`   from lavavu/petsc docker images to inherit respective compiled   functionality. Required subimage pip/apt packages are also installed   at this stage if they do not already exist in the base image. * Resultant base image should contain *only* runtime requireme
[…]
