+++ docs/development/docker/base/Dockerfile
@@ -1,151 +1,74 @@
 FROM debian:latest
 MAINTAINER https://github.com/underworldcode/
-
-# install things
-RUN apt-get update -qq && \
-    DEBIAN_FRONTEND=noninteractive apt-get install -yq --no-install-recommends \
+ENV LANG=C.UTF-8
+# Put this here in anticipation of virtualenvs
+ENV VIRTUAL_ENV=/opt/venv
+ENV PATH=${VIRTUAL_ENV}/bin:$PATH
+ENV PYTHONPATH=$PYTHONPATH:/usr/lib/python3.7/site-packages
+ENV LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib/:/usr/local/lib/x86_64-linux-gnu/
+# install runtime requirements
+RUN apt-get update -qq \
+&&  DEBIAN_FRONTEND=noninteractive apt-get install -yq --no-install-recommends \
         bash-completion \
-        build-essential \
-        git \
         python3-minimal \
-        python3-dev \
+        python3-virtualenv \
         python3-pip \
-        libxml2-dev \
-        xorg-dev \
-        ssh \
-        curl \
-        libfreetype6-dev \
-        libpng-dev \
-        libxft-dev \
-        xvfb \
-        freeglut3 \
-        freeglut3-dev \
-        libgl1-mesa-dri \
-        libgl1-mesa-glx \
         python3-tk \
-        rsync \
         vim \
         less \
-        xauth \
-        swig \
-        gdb-minimal \
-        python3-dbg \
-        cmake \
-        python3-setuptools \
-        wget \
-        gfortran  && \
-    apt-get clean && \
-    rm -rf /var/lib/apt/lists/*
+&&  apt-get clean \
+&&  rm -rf /var/lib/apt/lists/*
 
-ARG MPICH_VERSION="3.1.4"
-ARG MPICH_CONFIGURE_OPTIONS="--enable-fast=all,O3 --prefix=/usr"
-ARG MPICH_MAKE_OPTIONS="-j4"
-WORKDIR /tmp/mpich-build
-RUN wget http://www.mpich.org/static/downloads/${MPICH_VERSION}/mpich-${MPICH_VERSION}.tar.gz && \
-    tar xvzf mpich-${MPICH_VERSION}.tar.gz && \
-    cd mpich-${MPICH_VERSION}              && \
-    ./configure ${MPICH_CONFIGURE_OPTIONS} && \
-    make ${MPICH_MAKE_OPTIONS}             && \
-    make install                           && \
-    ldconfig                               && \
-    cd /tmp                                && \
-    rm -fr *
-
-ENV LANG=C.UTF-8
-# Install setuptools and wheel first, needed by plotly
-RUN pip3 install -U setuptools  && \
-    pip3 install -U wheel       && \
-    pip3 install --no-cache-dir packaging \
+RUN pip3 install -U setuptools  \
+&&  pip3 install --no-cache-dir \
+        packaging \
         appdirs \
         numpy \
         jupyter \
         plotly \
         matplotlib \
         pillow \
-        pyvirtualdisplay \
         ipython \
         ipyparallel \
         pint \
-        sphinx \
-        sphinx_rtd_theme \
-        sphinxcontrib-napoleon \
-        mock \
         scipy \ 
-        tabulate \
-        mpi4py \
-        scons \
-        lavavu
+        tabulate 
 
-WORKDIR /tmp/petsc-build
-ARG PETSC_VERSION="3.11.3"
-RUN wget http://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-${PETSC_VERSION}.tar.gz && \
[…]
