### 2019-11-01 #1797 @leouieda
@jmansour not that I'm aware of. You could always add the new affiliation and list the author as having both.
