### 2019-11-02 b1cdfcc
* Remove parallel jupyter extension from lavavu. * minimise dockerfiles.
