### 2019-11-02 10e1d31
Some tweaks for smaller images.
