### 2019-11-02 063c4de
* New UW2 Dockerfile for multistage etc. * Add git to base.
