### 2019-11-03 4fe4495
Further tweaks and commentary.
