+++ docs/development/docker/lavavu/Dockerfile
@@ -1,9 +1,12 @@
 FROM debian:latest as base_runtime
 ENV LANG=C.UTF-8
-# Put this here in anticipation of virtualenvs
+# Setup some things in anticipation of virtualenvs
 ENV VIRTUAL_ENV=/opt/venv
+# The following ensures that the venv takes precedence if available
 ENV PATH=${VIRTUAL_ENV}/bin:$PATH
-ENV PYTHONPATH=$PYTHONPATH:/usr/lib/python3.7/site-packages
+# The following ensures that system packages are visible to venv Python, and vice versa
+ENV PYTHONPATH=$PYTHONPATH:/usr/lib/python3.7/site-packages:${VIRTUAL_ENV}/lib/python3.7/site-packages
+# install runtime requirements
 RUN apt-get update -qq \
 &&  DEBIAN_FRONTEND=noninteractive apt-get install -yq --no-install-recommends \
         python3-minimal \
@@ -27,7 +30,6 @@ RUN apt-get update -qq \
         libllvm7 \
 &&  apt-get clean \
 &&  rm -rf /var/lib/apt/lists/*
-# Install setuptools and wheel first, needed by plotly
 RUN pip3 install -U setuptools  \
 &&  pip3 install --no-cache-dir \
         packaging \
@@ -37,10 +39,8 @@ RUN pip3 install -U setuptools  \
         pillow \
         ipython
 
-# install build requirements
 FROM base_runtime AS build_base
-# create a virtualenv to put new python modules
-RUN python3 -m virtualenv --python=/usr/bin/python3 ${VIRTUAL_ENV} 
+# install build requirements
 RUN apt-get update -qq 
 RUN DEBIAN_FRONTEND=noninteractive apt-get install -yq --no-install-recommends \
         build-essential \
@@ -74,6 +74,8 @@ ENV LV_LIB_DIRS=/usr/local/lib/x86_64-linux-gnu/
 ENV LV_INC_DIRS=/usr/local/include/
 ENV LV_OSMESA=1
 ENV LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib/x86_64-linux-gnu/
+# create a virtualenv to put new python modules
+RUN python3 -m virtualenv --python=/usr/bin/python3 ${VIRTUAL_ENV} 
 RUN pip3 install --no-cache-dir lavavu
 
 FROM base_runtime AS minimal
@@ -82,10 +84,11 @@ COPY --from=build_base /usr/local /usr/local
 ENV LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib/x86_64-linux-gnu/
 # Record Python packages, but only record system packages! 
 # Not venv packages, which will be copied directly in.
-RUN /usr/bin/pip3 freeze >/opt/requirements.txt
+RUN PYTHONPATH= /usr/bin/pip3 freeze >/opt/requirements.txt
 # Record manually install apt packages.
 RUN apt-mark showmanual >/opt/installed.txt
 
+# Add user environment
 FROM minimal
 EXPOSE 8888
 ADD https://github.com/krallin/tini/releases/download/v0.18.0/tini /tini
