### 2019-11-03 bdfe27b
Further tweaks for new Docker builds.
