### 2019-11-03 8841740
Update `CHANGES.md` to reflect docker changes.
