### 2019-11-07 8bf74aa
Update petsc Dockerfile to point to installed petsc, as petsc4py installs its own petsc (though it shouldn't).
