### 2019-11-07 4bfff4d
V2.8 release (#433)
* uw-2.8 should use the master branch of UWGeo

* improving jenkins container usage - as per dev branch

* Tether lavavu for 2.8 compatibility.

* Update README.md

* Update __init__.py

Increment version number to 2.8.3b

* Update CHANGES.md

* Update CHANGES.md

* Update jenkinsfile

Docker can't deal with upper case image name! 
So I'm naming the temporary image the jenkins builds after env.GIT_COMMIT. GIT_COMMIT is the changeset sha from the repo.
