+++ CHANGES.md
@@ -1,6 +1,16 @@
 CHANGES: Underworld2
 =======================
 
+Release 2.8.3b [2019-10-29]
+---------------------------
+Docker:
+* Revert to a compatible version of lavavu (v1.4.3).
+* Revert to a stable version of UWGeodynamics.
+
+Documentation:
+* Update `README.md` to with new Jenkins url. 
+
+
 Release 2.8.2b [2019-10-28]
 ---------------------------
 Docker:
