### 2019-11-11 b3f3bc8
Fixes #428.
Adding author to paper.
