### 2019-11-11 underworldcode/underworld2#428 @jmansour
Authorship finalised.

openjournals/joss-reviews/issues/1797

Thanks @gassmoeller.
