+++ underworld/libUnderworld/StgFEM/Utils/src/SemiLagrangianIntegrator.c
@@ -395,6 +395,12 @@ Bool PeriodicUpdate( double* pos, double* min, double* max, unsigned dim, Bool i
 Bool BicubicInterpolatorNew( FeVariable* feVariable, FeVariable* stencilField, double* position, unsigned* sizes, double* result ) {
   /* Calculated the BicubicInterpolation of the feVariable at position
    *
+   * Input Args:
+   *   feVariable: the field to be interpolated at `position`.
+   *   stencilField: the field of initial spline stencils for each node.
+   *   position:   the position of interpolatation.
+   *   sizes:      memory chunk of size = sizeof(double)*dim.
+   *   results:    the interpolated value.
    *
    *
    * Returns Values:
@@ -435,7 +441,7 @@ Bool BicubicInterpolatorNew( FeVariable* feVariable, FeVariable* stencilField, d
          for( x_i = 0; x_i < 4; x_i++ ) {
             gNode_I = ijk[0] + x_i + ( ijk[1] + y_i ) * sizes[0];
             if( !Mesh_GlobalToDomain( feMesh, MT_VERTEX, gNode_I, &lNode_I ) ){
-              printf("Error in BicubicInterpolator(), trying to build an interpolation to position (%g, %g) using node %d, a non domain node, in interpolator.\n", position[0], position[1], gNode_I);
+              printf("Error in %s, trying to build an interpolation to position (%g, %g) using node %d, a non domain node, in interpolator.\n", __func__, position[0], position[1], gNode_I);
               abort();
             }
             else
@@ -448,7 +454,7 @@ Bool BicubicInterpolatorNew( FeVariable* feVariable, FeVariable* stencilField, d
             for( x_i = 0; x_i < 4; x_i++ ) {
                gNode_I = ijk[0] + x_i + ( ijk[1] + y_i ) * sizes[0] + ( ijk[2] + z_i ) * sizes[0] * sizes[1];
                if( !Mesh_GlobalToDomain( feMesh, MT_VERTEX, gNode_I, &lNode_I ) ) {
-                 printf("Error in BicubicInterpolator(), trying to build an interpolation to position (%g, %g, %g) using node %d, a non domain node, in interpolator.\n", position[0], position[1], position[2], gNode_I);
+                 printf("Error in %s, trying to build an interpolation to position (%g, %g, %g) using node %d, a non domain node, in interpolator.\n", __func__, position[0], position[1], position[2], gNode_I);
                  abort();
                }
                else
@@ -494,6 +500,156 @@ Bool BicubicInterpolatorNew( FeVariable* feVariable, FeVariable* stencilField, d
    }
 
    return True;
+}
+
+
+
+Bool SemiLagrangianIntegrator_PointsAreClose( double* p1, double* p2, int dim, double rtol, double atol ) {
+  /* check if two points are within rtol (relative tolerance) 
+   * or atol (absolute tolerance)
+   * 
+   * Intput Parameters:
+   *  p1   : point 1
+   *  p2   : point 2
+   *  dim  : the dimensions of points
+   *  rtol : the relative tolerance
+   *  atol : the absolute tolerance
+   *
+   * Return Values:
+   *  True:  if points are close
+   *  False: if points are not close 
+   */
+
+  double disp[3], p2_norm, length;
+
+  StGermain_VectorSubtraction(disp, p1, p2, dim); // displacement vector
+  length  = StGermain_VectorMagnitude(disp, dim); // length of vector
[…]
