### 2019-11-13 3338cc8
Jgiordani/otho slcn stencil (#434)
This change replaces the (error prone) code to find where to build a (2D) 4x4 patch for spline creation.

Significant changes are:

The new patch is calculated using a node based method for finding the ijk starting point for every domain node.
The ijk start point is calculated once only, stored and reused.
The node based method is better than the previous algorithm because it uses node indices to determine if it's near an edge boundary instead of "domain"/global coordinates which were bad because floating point approx. and off domain edge cases occurred.

Ass
[…]
