### 2019-11-20 bed552f
'long' int swarmvariable implementation was incomplete, and as such `evaluate()` operations were not functioning.
* Add required implementation.
* Tidy swarm variable test, and include `evaluate()` sanity checks.
