### 2019-11-20 03053b6
Update `CHANGES.md` for `long` SwarmVariable changes.
