### 2019-11-21 d53281f
Prelim installation guide for Gadi. Note that petsc modules do not exist yet, so you'll have to build your own.  Same for swig, and it better be swig3 as swig4 is problematic.
