### 2019-12-05 #1797 @danielskatz
👋 @gabersyd - How is this going?

31 October:
> @leouieda, I need one more week to give a complete feedback.
