### 2020-01-02 #1797 @gabersyd
A comment: the installation folder (underworld2/docs/install_guides/) contains examples only for large HPC facilities. It would be useful to offer one also for a smaller machine, so that students could try using it in a smaller machine like a powerful desktop or a small cluster that they might want to access.

I tried to follow some existing instructions like the one for Caltech fram on our Queen Bee cluster in Louisiana, but I was stuck as the system asked me to install "scons". I did not see this requirement in the instructions.
