### 2020-01-02 #1797 @gabersyd
I finished the review. I just have an issue as I struggled with
dependencies when trying to install it on our largest cluster here in
Louisiana. Maybe we can look at it together, as the dependencies were not
all clear to me when I tried.

Thank you
Gabriele

Gabriele Morra
<http://scholar.google.com/citations?sortby=pubdate&hl=en&user=DDuTppAAAAAJ&view_op=list_works>
Associate Professor <https://physics.louisiana.edu/about-us/faculty-staff>
Department of Physics <http://physics.louisiana.edu> and School of
Geosciences <http://geos.louisiana.edu>
University of Louisiana at Lafayette <http://www.louisiana.edu>
Editor in Chief of *Artificial Intelligence in Geosciences
<http://www.keaipublishing.com/en/journals/artificial-intelligence-in-geosciences/>*

On Thu, Dec 5, 2019 at 4:52 AM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @gabersyd <https://github.com/gabersyd> - How is this going?
>
> 31 October:
>
> @leouieda <https://github.com/leouieda>, I need one more week to give a
> complete feedback.
>
