### 2020-01-04 #1797 @jmansour
Thanks @gabersyd!

Note that there are updated installation instructions on the development (and joss) branch. Apologies, I thought I'd mentioned that earlier, but checking this thread I can't see mention of it. Here's the corresponding ReadTheDocs page:

https://underworld2.readthedocs.io/en/development/Installation.html

I think that should give reasonable coverage for native machine installation. More than happy to have a closer look at installation on Queen Bee so see where things are going awry.
