### 2020-01-06 #1797 @gabersyd
Hi John, thank you for your answer. I have tried to follow all the
instructions and got stuck again into the installation of scons. I am not
familiar with this library. I looker at scons.org and it seems that they
have mostly developed for
Python 2.*, which is quite obsolete. I tried to install it manually on the
cluster, and I have gotten:

UnsatisfiableError: The following specifications were found to be in
conflict:

  - python 3.5*

  - scons -> python 2.7*

Use "conda info <package>" to see the dependencies for each package.

Is scons so necessary for the installation of Underworld, that there is no
way to get around it? Very few Python users today use Python 2.*, now that
it arrive to Python 3.8.

Do you suggest me just to use Docker? I am concerned whether the
performance on the cluster will be the same if I use Docker instead of a
native installation. Is it just the same?

Gabriele Morra
<http://scholar.google.com/citations?sortby=pubdate&hl=en&user=DDuTppAAAAAJ&view_op=list_works>
Associate Professor <https://physics.louisiana.edu/about-us/faculty-staff>
Department of Physics <http://physics.louisiana.edu> and School of
Geosciences <http://geos.louisiana.edu>
University of Louisiana at Lafayette <http://www.louisiana.edu>
Editor in Chief of *Artificial Intelligence in Geosciences
<http://www.keaipublishing.com/en/journals/artificial-intelligence-in-geosciences/>*

On Sat, Jan 4, 2020 at 5:49 PM John Mansour <notifications@github.com>
wrote:

> Thanks @gabersyd <https://github.com/gabersyd>!
>
> Note that there are updated installation instructions on the development
> (and joss) branch. Apologies, I thought I'd mentioned that earlier, but
> checking this thread I can't see mention of it. Here's the corresponding
> ReadTheDocs page:
>
> https://underworld2.readthedocs.io/en/development/Installation.html
>
> I think that should give reasonable coverage for native machine
> installation. More than happy to have a closer look at installation on
> Queen Bee so see where things are going awry.
>
