### 2020-01-07 #1797 @jmansour
Unfortunately `scons` is the package we use to build Underworld, so it is necessary.  There should be no issue with using it with Py3, but I don't have experience with `conda`.  Typically, I just use `pip install scons` (or perhaps `pip install --user scons` for local install).   Actually, you should simply be able to download it from the `scons` project page, untar, and then update your `PATH` environment variable so that it can locate the `scons` Python executable (in the untarred `script` directory).

https://scons.org/pages/download.html

Docker is low (Windows/OSX) or no (Linux) overhead. However, on cluster machines, there are caveats to usage, and it is generally better to use something like Singularity or Shifter which act as conduits to Docker environments in HPC environments.  It really depends on how the machine is being used and you might need to chat with the system admins to see what is the best way forward. Actually there are even performance advantages to using Docker on HPC for very large Python jobs, however, if not already setup, native build will probably be the path of least resistance. I'm happy to help if I can in either case.
