### 2020-01-09 #1797 @leouieda
@gabersyd thank you for your review! I see that all review items have been checked so I assume that you are happy to sign off on this.

SCons used to be Python 2.7 only but recent versions are Python 3.5+ only. Installing from PyPI (using pip) will get you the latest version (3.1.2), which should work. You can also install with conda from [conda-forge](https://github.com/conda-forge/scons-feedstock) where the latest version is 3.1.2 and should work in Python 3 (`conda install scons=3.1.2 -c conda-forge`). The issue is probably that the Anaconda package for SCons is outdated.
