### 2020-01-09 #1797 @leouieda
@jmansour could you please add the missing DOIs for the references outlined by Whedon above?
