### 2020-01-10 61c933f
Update bib with DOI info.
