+++ docs/joss/paper.bib
@@ -1,33 +1,36 @@
 @Article{MoresiEtAl2002,
-author="Moresi, L. and Dufour, F. and M{\"u}hlhaus, H.-B.",
-title="Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling",
-journal="Pure and Applied Geophysics",
-volume="159",
-number="10",
-pages="2335--2356",
-year="2002"
+	author="Moresi, L. and Dufour, F. and M{\"u}hlhaus, H.-B.",
+	title="Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling",
+	journal="Pure and Applied Geophysics",
+	volume="159",
+	number="10",
+	pages="2335--2356",
+	year="2002",
+	doi="10.1007/s00024-002-8738-3"
 }
 
 @Article{MoresiEtAl2003,
-author="Moresi, L. and Dufour, F. and M{\"u}hlhaus, H.-B.",
-title="A Lagrangian integration point finite element method for large deformation modeling of viscoelastic geomaterials",
-journal="Journal of Computational Physics",
-volume="184",
-number="2",
-pages="476--497",
-year="2003"
+	author="Moresi, L. and Dufour, F. and M{\"u}hlhaus, H.-B.",
+	title="A Lagrangian integration point finite element method for large deformation modeling of viscoelastic geomaterials",
+	journal="Journal of Computational Physics",
+	volume="184",
+	number="2",
+	pages="476--497",
+	year="2003",
+	doi="10.1016/S0021-9991(02)00031-1"
 }
 
 
 @article{MoresiEtAl2007,
-author = "L. Moresi and S. Quenette and V. Lemiale and C. Mériaux and B. Appelbe and H.-B. Mühlhaus",
-title = "Computational approaches to studying non-linear dynamics of the crust and mantle ",
-journal = "Physics of the Earth and Planetary Interiors ",
-volume = "163",
-number = "1–4",
-pages = "69 - 82",
-year = "2007",
-note = "Computational Challenges in the Earth Sciences "
+	author = "L. Moresi and S. Quenette and V. Lemiale and C. Mériaux and B. Appelbe and H.-B. Mühlhaus",
+	title = "Computational approaches to studying non-linear dynamics of the crust and mantle",
+	journal = "Physics of the Earth and Planetary Interiors",
+	volume = "163",
+	number = "1–4",
+	pages = "69 - 82",
+	year = "2007",
+	note = "Computational Challenges in the Earth Sciences",
+	doi = "10.1016/j.pepi.2007.06.009"
 }
 
 @conference{Kluyver:2016aa,
@@ -37,7 +40,8 @@ note = "Computational Challenges in the Earth Sciences "
 	Organization = {IOS Press},
 	Pages = {87 - 90},
 	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
-	Year = {2016}
+	Year = {2016},
+	doi = {10.3233/978-1-61499-649-1-87}
 }
 
 @article{Beucher2019UWGeodynamics,
@@ -70,33 +74,34 @@ note = "Computational Challenges in the Earth Sciences "
 }
 
 @Article{BeallEtAl2018,
-author="Beall, A. and Moresi, Cooper C.M.",
-title="Formation of cratonic lithosphere during the initiation of plate tectonics",
-journal="Geology",
-volume="46",
-number="6",
-pages="487--490",
-year="2018"
+	author="Beall, A. and Moresi, Cooper C.M.",
+	title="Formation of cratonic lithosphere during the initiation of plate tectonics",
+	journal="Geology",
+	volume="46",
[…]
