### 2020-01-10 #1797 @jmansour
Ok, I think they should all be good now @leouieda.

And can confirm using `scons 3.12` with Py3 works fine. Thanks for input @leouieda.
