### 2020-01-13 1cd2a74
Editorial fixes for the JOSS paper
A few minor tweaks to the formatting of the paper and citations.
Also removed a sentence from the second-to-last paragraph that
was probably supposed to be a comment.
