### 2020-01-13 50deca6
Fix formatting for Mériaux reference
