+++ docs/joss/paper.bib
@@ -85,7 +85,7 @@
 }
 
 @article{MeriauxEtAl2018,
-	author = {Meriaux, Catherine and A May, Dave and Mansour, John and Chen, Zhihao and Kaluza, Owen},
+	author = {Catherine A. M{\'{e}}riaux and Dave A. May and John Mansour and Zhihao Chen and Owen Kaluza},	
 	year = {2018},
 	month = {07},
 	pages = {},
