### 2020-01-13 #1797 @leouieda
@jmansour thank you for the update. I submitted a PR with some editorial fixes (mostly references and citation style). Let me know once you review and merge these changes.
