### 2020-01-13 6d20464
Editorial fixes for the JOSS paper (#451)
* Editorial fixes for the JOSS paper

A few minor tweaks to the formatting of the paper and citations.
Also removed a sentence from the second-to-last paragraph that
was probably supposed to be a comment.
