+++ docs/joss/paper.bib
@@ -61,16 +61,18 @@
 	day = {11},
 }
 
-@misc{owen_kaluza_2019_3348451,
-  author       = {Owen Kaluza and
-                  Louis Moresi and
-                  John Mansour and
-                  David G Barnes},
-  title        = {OKaluza/LavaVu: v1.4.3},
-  month        = jul,
-  year         = 2019,
-  doi          = {10.5281/zenodo.3348451},
-  url          = {https://doi.org/10.5281/zenodo.3348451}
+@software{owen_kaluza_2019_3348451,
+	author       = {Owen Kaluza and
+					Louis Moresi and
+					John Mansour and
+					David G Barnes},
+	title        = {OKaluza/LavaVu: v1.4.3},
+	month        = jul,
+	year         = 2019,
+	publisher    = {Zenodo},
+	version      = {1.4.3},
+	doi          = {10.5281/zenodo.3348451},
+	url          = {https://doi.org/10.5281/zenodo.3348451}
 }
 
 @Article{BeallEtAl2018,
