### 2020-01-13 e2310ee
Further formatting tweak.
