### 2020-01-13 6605768
Force captilisation for LavaVu.
