+++ docs/joss/paper.bib
@@ -66,7 +66,7 @@
 					Louis Moresi and
 					John Mansour and
 					David G Barnes},
-	title        = {OKaluza/LavaVu: v1.4.3},
+	title        = {OKaluza/{L}ava{V}u: v1.4.3},
 	month        = jul,
 	year         = 2019,
 	publisher    = {Zenodo},
