### 2020-01-13 #1797 @jmansour
Thanks @leouieda, I've merged in those changes.  Also I've added some further minor formatting tweaks:
https://github.com/underworldcode/underworld2/commit/15052cf73429bedde7b564d5582b681f6a64507b
https://github.com/underworldcode/underworld2/commit/e2310ee64c57bbc30070810374d858e3bc075f3e
https://github.com/underworldcode/underworld2/commit/66057689d27f6870b0b446a8b44590efc3f0b2b3
