### 2020-01-17 #1797 @leouieda
@jmansour I'm happy to move on with acceptance of your paper :tada: Here are the steps you need to take now:

- Make a release with the latest changes from the review and **post the version number here**.
- Archive the release on Zenodo/figshare/etc and **post the DOI here**.
- Make sure that the title and author list (including ORCIDs) in the archive match the JOSS paper.
