### 2020-01-20 2a78756
Minor updates for these docs.
