### 2020-01-21 e4eb910
* Add zenodo metadata file * Updates for other docs.
