### 2020-01-28 5202e6c
Singularity info for Magnus usage.
