+++ docs/development/guidelines.md
@@ -55,6 +55,8 @@ once the following criterion are met:
    1. Any new functionality is sufficiently tested (via doctest or other means).
    2. Any new functionality is sufficiently documented.
    3. All tests pass.
+   4. Blog post update for significant changes. 
+   5. Update CHANGES.md.
    
 
 Bug reporting
