### 2020-01-30 bb78fcd
Update Magnus Singularity instructions to mention issue with home directory usage.
