### 2020-02-03 ebbe27f
* Update `Dockerfile` to set required file/directory ownership. * Update for Magnus usage instructions. * Add option to allow disabling the vis server as it is problematic   on Magnus/Singularity.
