+++ docs/install_guides/pawsey_magnus_container.md
@@ -17,15 +17,35 @@ user@magnus-1:~$ module load singularity
 ```
 
 2. Pull down the required image. Note that images are generally around 1GB, so you might 
-consider storing them somewhere in `/group` or `/scratch`.
+   consider storing them somewhere in `/group` or `/scratch`.
 ```shell
-user@magnus-1:~$ singularity pull /WHERE/TO/STORE/YOUR/IMAGE/IMAGENAME.sif docker://underworldcode/underworld2:2.8.1b
+user@magnus-1:~$ singularity pull /WHERE/TO/STORE/YOUR/IMAGE/IMAGENAME.sif docker://underworldcode/underworld2:2.9.0b
 ```
 
 3. Use the image. Note that this will normally be a command line in your
-queue submission script. Don't forget to also load the Singularity module.
+   queue submission script. Don't forget to also load the Singularity module.
+   Note also the setting of the environment variable `UW_VIS_PORT` which is 
+   required for correct vis operation.
 ```shell
-user@magnus-1:~$ srun --export=all singularity exec /WHERE/TO/STORE/YOUR/IMAGE/IMAGENAME.sif python3 YOURSCRIPT.py
+user@magnus-1:~$ export UW_VIS_PORT=0
+user@magnus-1:~$ srun singularity exec /WHERE/TO/STORE/YOUR/IMAGE/IMAGENAME.sif python3 YOURSCRIPT.py
 ```
 
+For interactive usage, you will start a Singularity shell:
+
+1. First start an interactive session with however many tasks you require:
+```shell
+user@magnus-1:~$ salloc --ntasks=16
+```
+
+2. Once your session has begun, load the Singularity module:
+```shell
+user@magnus-1:~$ module load singularity
+```
+
+3. Launch the Singularity shell. Note the `--pty` required for pseudo terminal mode.
+```shell
+user@magnus-1:~$ export UW_VIS_PORT=0
+user@magnus-1:~$ srun --pty  singularity shell /WHERE/TO/STORE/YOUR/IMAGE/IMAGENAME.sif
+```
 
