### 2020-02-03 2b488bd
* Majorish tidying of Sphinx docs generation. Much cleaner now. * Moved `matplotlib_inline()` to `utils` submodule, and associated changes. * Vis submodule wasn't being documented by Sphinx. Fixed.
