+++ underworld/visualisation/_glucifer.py
@@ -183,8 +183,6 @@ class Store(_stgermain.StgCompoundComponent):
             db = self._db.path
         if 'UW_VIS_PORT' in os.environ:
             kwargs['port'] = int(os.environ['UW_VIS_PORT'])
-
-        import pdb;pdb.set_trace()
         return lavavu.Viewer(cache=False, clearstep=True, database=db, timestep=self.step, *args, **kwargs)
 
     def _generate(self, figname, objects, props):
