### 2020-02-07 a1b2ebd
* Switch to `kiloyear` for better `pint` compatibility. * Switch version number for beta.
