### 2020-02-07 9a63324
* Do not test vis when `lavavu` isn't available. * Remove `lavavu` install from Gadi install script.
