### 2020-02-10 4ef4c40
Changes for PyPi: * Updates to setup.py * Add `MANIFEST.in` to specify extra files required. * Update release guidelines.
