### 2020-02-10 8ce2124
Set required Dockerfile
