### 2020-02-10 #1797 @jmansour
Hi @leouieda

Ok, we have a new release out. Version 2.9.0b:

https://github.com/underworldcode/underworld2/releases/tag/v2.9.0b

Still waiting on Zenodo to generate the new corresponding DOI, but it should show up shortly here:
https://doi.org/10.5281/zenodo.1436039

Affiliations and ORCIDs have been confirmed. Note that as discussed earlier, one author has requested not to have an associated ORCID.

Let me know if you need anything further.

Thanks
