+++ underworld/_version.py
@@ -1,1 +1,1 @@
-__version__ = "2.9.0b"
+__version__ = "2.10.0-dev"
