### 2020-02-11 a0ed3a2
* Ensure Jupyter server proxy extension is enabled. * VOLUME workspace is now usable by joyvan * jupytext is enabled * Use pint 0.9 and UWGeo v29 * Update expected images as lavavu now renders colourbar positions slightly differently. * Update 07_Vis guide as scaling for labels is now handled differently. * Update top level Dockerfile to point to required tag. * Other minor fixes
