+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM underworldcode/underworld2:v2.9.0b
+FROM underworldcode/underworld2:2.9.1b
 
 # Set the UW_MACHINE env variable for metrics
 ENV UW_MACHINE binder
