### 2020-02-25 #1797 @leouieda
@jmansour apologies for the delay. Thank you for posting the DOI link and version number. I double checked the paper and everything looks good. The only minor thing I have to ask is that you update the title of the Zenodo archive of https://zenodo.org/record/3661465 to match the JOSS paper title. Let me know when you get a chance to do that and we'll move on to acceptance :tada:
