### 2020-02-26 24614f5
* Update Zenodo title to match JOSS. * Increment versions.
