+++ underworld/_version.py
@@ -1,1 +1,1 @@
-__version__ = "2.9.1b"
+__version__ = "2.9.2b"
