### 2020-02-26 #1797 @jmansour
Hi @leouieda

Am I able to update the title of an existing Zenodo archive?   In any case, I've created a new release (v2.9.2) with the updated title. Here's the corresponding Zenodo reference:
https://zenodo.org/record/3687399
