### 2020-03-07 #1797 @jmansour
Excellent result!

I'd also like to thank @gassmoeller & @gabersyd for the time they put in reviewing this and the valuable feedback they provided. 🙏🙏

And thanks to @leouieda & @Kevin-Mattheus-Moerman and the rest of the JOSS crew. The review process was straightforward and even fun. 🙂
